tasks/vision/classification/inceptionv4/requirements.txt
numpy==1.24.3
tensorflow==2.13.1
Pillow==10.3.0